requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
psycopg2-binary
boto3
httpx
